requirements.txt
aiofiles==0.6.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
APScheduler==3.9.1
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncprawcore==2.3.0
attrs==21.4.0
autopep8==1.6.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
Flask==2.1.1
frozenlist==1.3.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
multidict==6.0.2
outcome==1.1.0
praw==7.5.0
prawcore==2.3.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.35
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
twilio==7.8.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
update-checker==0.18.0
urllib3==1.26.9
websocket-client==1.3.2
Werkzeug==2.1.1
wsproto==1.1.0
yarl==1.7.2
